docs: refine README title and badge placement
In this commit, the README's title format has been revised to enhance
clarity and usability. The modification involves displaying the badge
subsequent to the title, rather than integrating it within the title.
This change addresses potential confusion experienced by some users due
to the previous format where the badge was part of the title.