feat: Implement Custom Footer
This commit replaces the default footer with a custom message. Currently,
it utilizes a fork of the 'jekyll-build-pages' action
(https://github.com/actions/jekyll-build-pages) due to an unmerged pull
request (https://github.com/pages-themes/primer/pull/85). Once the pull
request is merged into the upstream, this change can be reverted.